+++ MiniMap.js
@@ -98,7 +98,6 @@ define(function (require, exports, module) {
                 }
             }
         });
-
     }
 
     exports.init = init;